Refactor: Introduce main menu and move game logic to testRoom.ts
- Moved core game simulation logic from main.ts to a new file src/game/utils/testRoom.ts.
- The logic in testRoom.ts is encapsulated within an exported startTestRoom function.
- Updated main.ts to display a simple main menu with a 'Start Test Room' button.
- Clicking the button hides the menu and invokes startTestRoom to launch the game environment.
- Adjusted import paths and state management within testRoom.ts to reflect the new structure and ensure correct operation.
